[Unchained] Add populateTriplestore as CLI arg (#367)
Temporary solution, since I don't have a better short-term idea at the
moment.
Runs a population of the external triplestore (Virtuoso) from RDFLog if
the commandline argument `populateTriplestore` is given. Based on the
logic that was original in `tools/populateTriplestore.scala`.
